Set permissions to container uid and group
docker cp has a bug where it uses the host machine's uid and group
for copied files and directories even though the copied files and
directories should use the container's user's uid and group. To fix
this, we run chown as root on the copied files and directories to
get the correct permissions.